requirements.txt
praw
discord.py
numpy
pandas
rasa
python-dotenv